check workspace tags in schedules
